Add testsuite tests count to fix Bitbucket pipelines ignoring output
Without this tag, Bitbucket pipelines will pick up the file, but it will 
not visualize it in its interface.

Strangely enough, this does work if you also have other test reports in 
the expected format present, at which point it probably merges it and 
everything is displayed just fine. In other words, this only broke if 
there was just one test report present and it was this one.